eismoinfo_scraper/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-completion==0.5.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crayons==0.4.0
cron-descriptor==1.4.5
Django==5.1.1
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-prometheus>=2.3.1
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.8
flake8==7.1.1
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.4.1
lxml==5.3.0
MarkupSafe==3.0.1
mccabe==0.7.0
multidict==6.1.0
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.47
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycycle==0.0.8
pyflakes==3.2.0
pytest==8.3.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
shellingham==1.5.4
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stubs==1.0.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.17.1